Feat: Quest 수정 시작

+++ backend/src/main/java/com/ssafy/db/entity/User/UserProfile.java
@@ -25,8 +25,8 @@ public class UserProfile {
     private int rest_point;
 
     // 위도
-    private int latitude;
+    private double latitude;
 
     // 경도
-    private int longitude;
+    private double longitude;
 }
